test: configure test project to not run twice
Some tests had been configured in a way that they had both the extension
as well as an explicit call to the plugin. Which resulted the plugin
running twice, one time with the default configuration. Another time
with the explicit configuration.

This results in the RPM file being created twice. But in the case of the
"forcerelease" test, with too different output file names.

This change configures the test project correctly, so that there is no
explicit invocation of the plugin, only the configuration of the
implicit step, triggered by enabling the plugin as an extension.

closes #63