test: update expected checksums
This is necessary as the new version does add new header fields, and so
the binary output, and thus the checksum, changes.

As the tool (rpm builder) is considered a part of the build process,
changing the tool (version) might also change the outcome of the build.

If you want to use the new tool, you need to update the expectation of
the outcome. If you want to keep the expected outcome, you also need
to use the old version of the tool.