feat: Default file digest algorithm to SHA-256 and allow to override
This change allows to configure the digest algorithm used for files
inside the RPM file. Previously this defaulted to MD5. It is now also
possible to override this.

BREAKING CHANGE: By default, this will now create RPMs with a file
digest algorithm of SHA-256. It is possible to override this back to
MD5.

Closes: #70