test: update the MD5 checksums of the reproducible build tests
As the content of the RPM file changes, due to switching the default
for file digests from MD5 to SHA-256, also the MD5 checksum of the
generated file will be different.

The reason we still use MD5 for this test is that we are not interested
in the actual checksum or algorithm, but want to ensure that the RPM
gets created in a reproducible way.

In theory it should be possible to set the file digest algorithm to MD5
and get the exact same (reproducible) file as with version 1.9. However,
we don't make any promises that the build stays reproducible between
minor version changes, and so there is no real reason to test this. The
tool version is considered part of the factors that influence the output
of the build.